[#1584] Cleanup and restructure code
