+++ iceoryx2/src/service/mod.rs
@@ -259,7 +259,6 @@ use core::time::Duration;
 
 use alloc::format;
 use alloc::string::String as CoreString;
-use alloc::string::ToString;
 use alloc::sync::Arc;
 use alloc::vec;
 use alloc::vec::Vec;
@@ -268,17 +267,28 @@ use iceoryx2_bb_elementary_traits::testing::abandonable::Abandonable;
 use iceoryx2_bb_posix::file::AccessMode;
 
 use crate::config;
-use crate::constants::MAX_TYPE_NAME_LENGTH;
 use crate::identifiers::UniqueServiceId;
-use crate::identifiers::{UniqueNodeId, UniquePortId};
 use crate::node::{NodeListFailure, NodeState, SharedNode};
-use crate::service::config_scheme::{dynamic_config_storage_config, port_tag_config};
+use crate::service::config_scheme::dynamic_config_storage_config;
 use crate::service::dynamic_config::DynamicConfig;
 use crate::service::naming_scheme::dynamic_config_name;
 use crate::service::naming_scheme::static_config_name;
+use crate::service::stale_resource_cleanup::{
+    remove_additional_blackboard_resources,
+    remove_sender_and_receiver_connections_and_data_segment,
+    remove_sender_connection_and_data_segment,
+};
+use crate::service::stale_resource_cleanup::{remove_service_tag, remove_static_service_config};
 use crate::service::static_config::*;
-use config_scheme::service_tag_config;
-use iceoryx2_bb_container::semantic_string::SemanticString;
+use crate::{
+    identifiers::{UniqueNodeId, UniquePortId},
+    node::NodeBuilder,
+    port::{listener::remove_connection_of_listener, notifier::Notifier},
+    prelude::EventId,
+    service::stale_resource_cleanup::{remove_port_tag, remove_receiver_port_from_all_connections},
+};
+use builder::event::EventOpenError;
+use dynamic_config::PortCleanupAction;
 use iceoryx2_bb_elementary::CallbackProgression;
 use iceoryx2_cal::arc_sync_policy::ArcSyncPolicy;
 use iceoryx2_cal::dynamic_storage::{
@@ -296,7 +306,9 @@ use iceoryx2_cal::shared_memory::{SharedMemory, SharedMemoryForPoolAllocator};
 use iceoryx2_cal::shm_allocator::shm_bump_allocator::BumpAllocator;
 use iceoryx2_cal::static_storage::*;
 use iceoryx2_cal::zero_copy_connection::ZeroCopyConnection;
-use iceoryx2_log::{debug, error, fail, trace, warn};
+use iceoryx2_log::error;
+use iceoryx2_log::{debug, fail, trace, warn};
+use port_factory::PortFactory;
 use service_hash::ServiceHash;
 
 use self::dynamic_config::DeregisterNodeState;
@@ -357,20 +369,6 @@ impl core::fmt::Display for ServiceRemoveError {
 }
 
 impl core::error::Error for ServiceRemoveError {}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum ServiceRemoveTagError {
-    AlreadyRemoved,
-    InternalError,
-    InsufficientPermissions,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum PortRemoveTagError {
-    AlreadyRemoved,
-    InternalError,
-    InsufficientPermissions,
-}
 
 /// Failure that can be reported when the [`ServiceDetails`] are acquired with [`Service::details()`].
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -563,28 +561,7 @@ impl<S: Service, R: ServiceResource> Drop for ServiceState<S, R> {
 
 #[doc(hidden)]
 pub mod internal {
-    use builder::event::EventOpenError;
-    use dynamic_config::PortCleanupAction;
-    use iceoryx2_bb_container::string::*;
-    use iceoryx2_log::error;
-    use port_factory::PortFactory;
-
-    use crate::{
-        identifiers::{UniqueNodeId, UniquePortId},
-        node::NodeBuilder,
-        port::{listener::remove_connection_of_listener, notifier::Notifier},
-        prelude::EventId,
-        service::stale_resource_cleanup::{
-            remove_data_segment_of_port, remove_receiver_port_from_all_connections,
-            remove_sender_port_from_all_connections,
-        },
-    };
-
     use super::*;
-
-    #[derive(Debug)]
-    struct CleanupFailure;
-
     fn send_dead_node_signal<S: Service>(service_hash: &ServiceHash, config: &config::Config) {
         let origin = "send_dead_node_signal()";
 
@@ -657,97 +634,52 @@ pub mod internal {
         trace!(from origin, "Send dead node signal on service {}.", service_name);
     }
 
-    fn remove_sender_connection_and_data_segment<S: Service>(
-        id: u128,
-        config: &config::Config,
-        origin: &str,
-        port_name: &str,
-    ) -> Result<(), CleanupFailure> {
-        unsafe { remove_sender_port_from_all_connections::<S>(id, config) }.map_err(|e| {
-            debug!(from origin,
-                "Failed to remove the {} ({:?}) from all of its connections ({:?}).",
-                port_name, id, e);
-            CleanupFailure
-        })?;
-
-        unsafe { remove_data_segment_of_port::<S>(id, config) }.map_err(|e| {
-            debug!(from origin,
-                "Failed to remove the {} ({:?}) data segment ({:?}).",
-                port_name, id, e);
-            CleanupFailure
-        })?;
-
-        Ok(())
-    }
-
-    fn remove_sender_and_receiver_connections_and_data_segment<S: Service>(
-        id: u128,
-        config: &config::Config,
-        origin: &str,
-        port_name: &str,
-    ) -> Result<(), CleanupFailure> {
-        remove_sender_connection_and_data_segment::<S>(id, config, origin, port_name)?;
-        unsafe { remove_receiver_port_from_all_connections::<S>(id, config) }.map_err(|e| {
-            debug!(from origin,
-                    "Failed to remove the {} ({:?}) from all of its incoming connections ({:?}).",
-                    port_name, id, e);
-            CleanupFailure
-        })?;
-
-        Ok(())
-    }
-
-    fn remove_additional_blackboard_resources<S: Service>(
-        config: &config::Config,
-        blackboard_name: &FileName,
-        blackboard_payload_config: &<S::BlackboardPayload as NamedConceptMgmt>::Configuration,
-        blackboard_mgmt_name: &StaticString<MAX_TYPE_NAME_LENGTH>,
-        origin: &str,
-        msg: &str,
-    ) {
-        match unsafe {
-            <S::BlackboardPayload as NamedConceptMgmt>::remove_cfg(
-                blackboard_name,
-                blackboard_payload_config,
-            )
-        } {
-            Ok(true) => {
-                trace!(from origin, "Remove blackboard payload segment.");
-            }
-            _ => {
-                error!(from origin,
-                                  "{} since the blackboard payload segment cannot be removed - service seems to be in a corrupted state.", msg);
-            }
-        }
-
-        // u64 is just a placeholder needed for the DynamicStorageConfiguration; it is
-        // overwritten right below
-        let mut blackboard_mgmt_config =
-            crate::service::config_scheme::blackboard_mgmt_config::<S, u64>(config);
-        // Safe since the same type name is set when creating the BlackboardMgmt in
-        // Creator::create_impl so we can safely remove the concept.
-        unsafe {
-            <S::BlackboardMgmt<u64> as DynamicStorage<u64>>::__internal_set_type_name_in_config(
-                &mut blackboard_mgmt_config,
-                blackboard_mgmt_name.as_str(),
-            )
-        };
-        match unsafe {
-            <S::BlackboardMgmt<u64> as NamedConceptMgmt>::remove_cfg(
-                blackboard_name,
-                &blackboard_mgmt_config,
-            )
-        } {
-            Ok(true) => {
-                trace!(from origin, "Remove blackboard mgmt segment.");
-            }
-            _ => {
-                error!(from origin, "{} since the blackboard mgmt segment cannot be removed - service seems to be in a corrupted state.", msg);
-            }
-        }
-    }
-
     pub trait ServiceInternal<S: Service> {
+        /// # Safety
+        ///
+        /// * No other process shall using the service currently.
+        ///
+        #[doc(hidden)]
+        unsafe fn __internal_force_remove_service(
+            service_name: &ServiceName,
+            config: &config::Config,
+            messaging_pattern: MessagingPattern,
+        ) -> Result<bool, ServiceRemoveError> {
+            let origin = "Service::remove()";
+            let msg = format!("Unable to remove all resources of the service \"{service_name}\"");
+            let service_hash =
+                ServiceHash::new::<S::ServiceNameHasher>(service_name, messaging_pattern);
+            match read_static_service_config::<S>(config, &service_hash) {
+                Ok(Some(static_config)) => {
+                    unsafe {
+                        S::__internal_remove_service(
+                            &service_hash,
+                            static_config.unique_service_id(),
+                            config,
+                        )?
+                    };
+
+                    Ok(true)
+                }
+                Ok(None) => Ok(false),
+                Err(e) => {
+                    warn!(from origin, "{msg} since the static config could not be read. [{e:?}]");
+
+                    match unsafe { remove_static_service_config::<S>(config, &service_hash) } {
+                        Ok(v) => Ok(v),
+                        Err(NamedConceptRemoveError::InsufficientPermissions) => {
+                            fail!(from origin, with ServiceRemoveError::InsufficientPermissions,
+                                "{msg} due to insufficient permissions.");
+                        }
+                        Err(NamedConceptRemoveError::InternalError) => {
+                            fail!(from origin, with ServiceRemoveError::InternalError,
+                                "{msg} due to an internal error.");
+                        }
+                    }
+                }
+            }
+        }
+
         /// # Safety
         ///
         /// * No other process shall using the service currently.
@@ -1148,87 +1080,6 @@ pub trait Service: Debug + Sized + internal::ServiceInternal<Self> + Clone {
 
         Ok(())
     }
-
-    /// Removes a [`Service`] forcefully created under a given [`config::Config`].
-    /// Removing a [`Service`] in use is undefined behavior!
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use iceoryx2::prelude::*;
-    /// use iceoryx2::config::Config;
-    ///
-    /// # fn main() -> Result<(), Box<dyn core::error::Error>> {
-    /// ipc::Service::remove("My/Funky/ServiceName".into(),
-    ///                      Config::global_config(),
-    ///                      MessagingPattern::PublishSubscribe)?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    ///
-    /// # Safety
-    ///
-    /// * No other process or thread shall currently use this service.
-    ///
-    unsafe fn remove(
-        service_name: &ServiceName,
-        config: &config::Config,
-        messaging_pattern: MessagingPattern,
-    ) -> Result<bool, ServiceRemoveError> {
-        let origin = "Service::remove()";
-        let msg = format!("Unable to remove all resources of the service \"{service_name}\"");
-        let service_hash =
-            ServiceHash::new::<Self::ServiceNameHasher>(service_name, messaging_pattern);
-        match read_static_service_config::<Self>(config, &service_hash) {
-            Ok(Some(static_config)) => {
-                unsafe {
-                    Self::__internal_remove_service(
-                        &service_hash,
-                        static_config.unique_service_id(),
-                        config,
-                    )?
-                };
-
-                Ok(true)
-            }
-            Ok(None) => Ok(false),
-            Err(e) => {
-                warn!(from origin,
-                    "{msg} since the static config could not be read. [{e:?}]");
-
-                match unsafe { remove_static_service_config::<Self>(config, &service_hash) } {
-                    Ok(v) => Ok(v),
-                    Err(NamedConceptRemoveError::InsufficientPermissions) => {
-                        fail!(from origin, with ServiceRemoveError::InsufficientPermissions,
-                            "{msg} due to insufficient permissions.");
-                    }
-                    Err(NamedConceptRemoveError::InternalError) => {
-                        fail!(from origin, with ServiceRemoveError::InternalError,
-                            "{msg} due to an internal error.");
-                    }
-                }
-            }
-        }
-    }
-}
-
-pub(crate) unsafe fn remove_static_service_config<S: Service>(
-    config: &config::Config,
-    service_hash: &ServiceHash,
-) -> Result<bool, NamedConceptRemoveError> {
-    let msg = "Unable to remove static service config";
-    let origin = "Service::remove_static_service_config()";
-    let name = static_config_name(service_hash);
-    let static_storage_config = config_scheme::static_config_storage_config::<S>(config);
-
-    match unsafe {
-        <S::StaticStorage as NamedConceptMgmt>::remove_cfg(&name, &static_storage_config)
-    } {
-        Ok(v) => Ok(v),
-        Err(e) => {
-            fail!(from origin, with e, "{msg} due to ({:?}).", e);
-        }
-    }
 }
 
 #[doc(hidden)]
@@ -1354,75 +1205,5 @@ fn open_dynamic_config<S: Service>(
                 fail!(from origin, with ServiceDetailsError::InternalError,
                     "{} due to an internal failure while opening the services dynamic config.", msg);
             }
-    }
-}
-
-pub(crate) fn remove_service_tag<S: Service>(
-    node_id: &UniqueNodeId,
-    service_hash: &ServiceHash,
-    config: &config::Config,
-) -> Result<(), ServiceRemoveTagError> {
-    let origin = format!(
-        "remove_service_tag<{}>({:?}, service_hash: {:?})",
-        core::any::type_name::<S>(),
-        node_id,
-        service_hash
-    );
-
-    match unsafe {
-        <S::StaticStorage as NamedConceptMgmt>::remove_cfg(
-            &service_hash.0.into(),
-            &service_tag_config::<S>(config, node_id),
-        )
-    } {
-        Ok(true) => Ok(()),
-        Ok(false) => {
-            fail!(from origin, with ServiceRemoveTagError::AlreadyRemoved,
-                    "The service's tag for the node was already removed. This may indicate a corrupted system!");
-        }
-        Err(NamedConceptRemoveError::InternalError) => {
-            fail!(from origin, with ServiceRemoveTagError::InternalError,
-                "Unable to remove the service's tag for the node due to an internal error.");
-        }
-        Err(NamedConceptRemoveError::InsufficientPermissions) => {
-            fail!(from origin, with ServiceRemoveTagError::InsufficientPermissions,
-                "Unable to remove the service's tag for the node due to insufficient permissions.");
-        }
-    }
-}
-
-pub(crate) fn remove_port_tag<S: Service>(
-    node_id: &UniqueNodeId,
-    port_id: &UniquePortId,
-    config: &config::Config,
-) -> Result<(), PortRemoveTagError> {
-    let origin = format!(
-        "remove_port_tag<{}>({:?}, port_id: {:?})",
-        core::any::type_name::<S>(),
-        node_id,
-        port_id
-    );
-    let name = FileName::new(port_id.value().to_string().as_bytes())
-        .expect("A number is always a valid file name.");
-
-    match unsafe {
-        <S::StaticStorage as NamedConceptMgmt>::remove_cfg(
-            &name,
-            &port_tag_config::<S>(config, node_id),
-        )
-    } {
-        Ok(true) => Ok(()),
-        Ok(false) => {
-            fail!(from origin, with PortRemoveTagError::AlreadyRemoved,
-                    "The port's tag for the node was already removed. This may indicate a corrupted system!");
-        }
-        Err(NamedConceptRemoveError::InternalError) => {
-            fail!(from origin, with PortRemoveTagError::InternalError,
-                "Unable to remove the port's tag for the node due to an internal error.");
-        }
-        Err(NamedConceptRemoveError::InsufficientPermissions) => {
-            fail!(from origin, with PortRemoveTagError::InsufficientPermissions,
-                "Unable to remove the port's tag for the node due to insufficient permissions.");
-        }
     }
 }
